[platform-sdk] Enable cross-compiliation for Android

+++ src/airmap/date_time.cpp
@@ -141,35 +141,35 @@ Microseconds DateTime::time_of_day() const {
 
 Hours hours(int64_t raw) {
   Hours hours;
-  hours.impl->time_duration = boost::posix_time::hours{raw};
+  hours.impl->time_duration = boost::posix_time::hours{static_cast<long>(raw)};
 
   return hours;
 }
 
 Minutes minutes(int64_t raw) {
   Minutes minutes;
-  minutes.impl->time_duration = boost::posix_time::minutes{raw};
+  minutes.impl->time_duration = boost::posix_time::minutes{static_cast<long>(raw)};
 
   return minutes;
 }
 
 Seconds seconds(int64_t raw) {
   Seconds seconds;
-  seconds.impl->time_duration = boost::posix_time::seconds{raw};
+  seconds.impl->time_duration = boost::posix_time::seconds{static_cast<long>(raw)};
 
   return seconds;
 }
 
 Milliseconds milliseconds(int64_t raw) {
   Milliseconds milliseconds;
-  milliseconds.impl->time_duration = boost::posix_time::milliseconds{raw};
+  milliseconds.impl->time_duration = boost::posix_time::milliseconds{static_cast<long>(raw)};
 
   return milliseconds;
 }
 
 Microseconds microseconds(int64_t raw) {
   Microseconds microseconds;
-  microseconds.impl->time_duration = boost::posix_time::microseconds{raw};
+  microseconds.impl->time_duration = boost::posix_time::microseconds{static_cast<long>(raw)};
 
   return microseconds;
 }
